+++ lib/flutter_async.dart
@@ -2,6 +2,8 @@
 library flutter_async;
 
 export 'src/configs/async_config.dart';
+export 'src/extensions/element.dart';
+export 'src/extensions/future.dart';
 export 'src/widgets/async/async.dart';
 // AsyncBuilder
 export 'src/widgets/async_builder/async_builder.dart';
@@ -9,5 +11,6 @@ export 'src/widgets/async_builder/async_builder.dart';
 export 'src/widgets/async_button/async_button_builder.dart';
 export 'src/widgets/async_button/async_button_style_button.dart';
 export 'src/widgets/async_button/async_floating_action_button.dart';
+export 'src/widgets/async_button/async_icon_button.dart';
 // AsyncIndicator
 export 'src/widgets/async_indicator/async_indicator.dart';